# LLT-5122 Merge tcp-udp incoming handler