--- a/crates/telio-model/src/config.rs
+++ b/crates/telio-model/src/config.rs
@@ -41,6 +41,10 @@ pub struct Peer {
     pub is_local: bool,
     /// Flag to control whether the peer allows incoming connections
     pub allow_incoming_connections: bool,
+    /// Flag to control whether the Node allows routing through
+    pub allow_routing: bool,
+    /// Flag to control whether the Node allows incoming local area access
+    pub allow_local_area_access: bool,
     #[serde(default)]
     /// Flag to control whether the peer allows incoming files
     pub allow_peer_send_files: bool,
@@ -355,6 +359,8 @@ mod tests {
                   "is_local": true,
                   "user_email": "alice@example.com",
                   "allow_incoming_connections": true,
+                  "allow_routing": true,
+                  "allow_local_area_access": true,
                   "allow_peer_send_files": true,
                   "peer_allows_traffic_routing": false,
                   "allow_peer_traffic_routing": true
@@ -462,6 +468,8 @@ mod tests {
                     },
                     is_local: true,
                     allow_incoming_connections: true,
+                    allow_routing: true,
+                    allow_local_area_access: true,
                     allow_peer_send_files: true,
                 },
                 Peer {
@@ -476,6 +484,8 @@ mod tests {
                     },
                     is_local: false,
                     allow_incoming_connections: false,
+                    allow_routing: false,
+                    allow_local_area_access: false,
                     allow_peer_send_files: false,
                 },
             ]),