--- a/crates/telio-traversal/Cargo.toml
+++ b/crates/telio-traversal/Cargo.toml
@@ -10,7 +10,6 @@ publish = false
 bytecodec = "0.4.15"
 enum-map = "2.5.0"
 http = "0.2.8"
-if-addrs = "0.12.0"
 igd = { version = "0.12.1", features = ["aio"] }
 sm = "0.9.0"
 stun_codec = "0.1.13"
@@ -20,6 +19,7 @@ base64.workspace = true
 boringtun.workspace = true
 futures.workspace = true
 httparse.workspace = true
+if-addrs.workspace = true
 ipnet.workspace = true
 lazy_static.workspace = true
 tracing.workspace = true