--- a/crates/telio-utils/Cargo.toml
+++ b/crates/telio-utils/Cargo.toml
@@ -12,6 +12,8 @@ sn_fake_clock = ["dep:sn_fake_clock"]
 [dependencies]
 futures.workspace = true
 hashlink.workspace = true
+if-addrs.workspace = true
+ipnet.workspace = true
 mockall = { workspace = true, optional = true }
 parking_lot.workspace = true
 rustc-hash.workspace = true