### Problem

In order to change the firewall logic of handling incoming connections, it is better if we have a combined method to handle all the connections. It would lead to easier editing and handling policies.

P.S. - ICMP is not yet merged into the combined handler as it is more complex than merging tcp and udp.
### Solution
Merge UDP and TCP incoming connection handler.


### :ballot_box_with_check: Definition of Done checklist
- [x] Commit history is clean ([requirements](../blob/main/docs/git_commit_messages_requirements.md))
- [x] README.md is updated
- [x] Functionality is covered by unit or integration tests