@Hasan6979 2024-04-30
### Firewall benchmarks
## Before
process inbound tcp packet - rejected/peers: 64, ip: v4
                        time:   [3.3776 ms 3.3860 ms 3.3952 ms]
                        thrpt:  [29.454 Melem/s 29.534 Melem/s 29.607 Melem/s]
                 change:
                        time:   [+0.3332% +0.6447% +0.9862%] (p = 0.00 < 0.05)
                        thrpt:  [-0.9765% -0.6406% -0.3320%]
                        Change within noise threshold.
Found 18 outliers among 100 measurements (18.00%)
  18 (18.00%) high severe
process inbound tcp packet - rejected/peers: 64, ip: v6
                        time:   [4.2041 ms 4.2145 ms 4.2258 ms]
                        thrpt:  [23.664 Melem/s 23.728 Melem/s 23.787 Melem/s]
                 change:
                        time:   [-11.723% -4.7234% -0.4134%] (p = 0.17 > 0.05)
                        thrpt:  [+0.4151% +4.9575% +13.280%]
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

Benchmarking process inbound tcp packet - accepted because of peer/peers: 64, ip: v4: Collecting 100 samples in estimated 1.0464 s (
process inbound tcp packet - accepted because of peer/peers: 64, ip: v4
                        time:   [1.3474 ms 1.3530 ms 1.3588 ms]
                        thrpt:  [73.595 Melem/s 73.909 Melem/s 74.215 Melem/s]
                 change:
                        time:   [-0.8129% +0.2094% +1.0827%] (p = 0.69 > 0.05)
                        thrpt:  [-1.0711% -0.2089% +0.8196%]
                        No change in performance detected.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
Benchmarking process inbound tcp packet - accepted because of peer/peers: 64, ip: v6: Collecting 100 samples in estimated 1.0821 s (
process inbound tcp packet - accepted because of peer/peers: 64, ip: v6
                        time:   [1.3334 ms 1.3395 ms 1.3463 ms]
                        thrpt:  [74.280 Melem/s 74.653 Melem/s 74.997 Melem/s]
                 change:
                        time:   [-3.1520% -0.6578% +0.9438%] (p = 0.67 > 0.05)
                        thrpt:  [-0.9350% +0.6622% +3.2546%]
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

Benchmarking process inbound tcp packet - accepted because of port/peers: 64, ip: v4: Collecting 100 samples in estimated 1.2451 s (
process inbound tcp packet - accepted because of port/peers: 64, ip: v4
                        time:   [5.9626 ms 6.0834 ms 6.3015 ms]
                        thrpt:  [15.869 Melem/s 16.438 Melem/s 16.771 Melem/s]
                 change:
                        time:   [+19.944% +22.459% +27.158%] (p = 0.00 < 0.05)
                        thrpt:  [-21.358% -18.340% -16.628%]
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
Benchmarking process inbound tcp packet - accepted because of port/peers: 64, ip: v6: Collecting 100 samples in estimated 1.4327 s (
process inbound tcp packet - accepted because of port/peers: 64, ip: v6
                        time:   [7.1427 ms 7.1774 ms 7.2191 ms]
                        thrpt:  [13.852 Melem/s 13.933 Melem/s 14.000 Melem/s]
                 change:
                        time:   [+19.997% +20.712% +21.435%] (p = 0.00 < 0.05)
                        thrpt:  [-17.651% -17.158% -16.665%]
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

process outbound tcp packet - SYN/peers: 64, ip: v4
                        time:   [4.1680 ms 4.1845 ms 4.2024 ms]
                        thrpt:  [23.796 Melem/s 23.898 Melem/s 23.992 Melem/s]
                 change:
                        time:   [-1.6452% -1.1452% -0.6024%] (p = 0.00 < 0.05)
                        thrpt:  [+0.6061% +1.1584% +1.6728%]
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
process outbound tcp packet - SYN/peers: 64, ip: v6
                        time:   [5.1051 ms 5.1162 ms 5.1277 ms]
                        thrpt:  [19.502 Melem/s 19.546 Melem/s 19.588 Melem/s]
                 change:
                        time:   [-3.0999% -2.4163% -1.8377%] (p = 0.00 < 0.05)
                        thrpt:  [+1.8722% +2.4761% +3.1990%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

process outbound tcp packet - other/peers: 64, ip: v4
                        time:   [2.6244 ms 2.6327 ms 2.6428 ms]
                        thrpt:  [37.839 Melem/s 37.984 Melem/s 38.104 Melem/s]
                 change:
                        time:   [-1.4904% -0.9603% -0.4530%] (p = 0.00 < 0.05)
                        thrpt:  [+0.4550% +0.9696% +1.5130%]
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) high mild
  11 (11.00%) high severe
process outbound tcp packet - other/peers: 64, ip: v6
                        time:   [3.1686 ms 3.1746 ms 3.1809 ms]
                        thrpt:  [31.438 Melem/s 31.501 Melem/s 31.559 Melem/s]
                 change:
                        time:   [-9.8178% -4.9871% -2.0450%] (p = 0.01 < 0.05)
                        thrpt:  [+2.0877% +5.2489% +10.887%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

Benchmarking process outbound tcp packet - accepted because of peer/peers: 64, ip: v4: Collecting 100 samples in estimated 1.0233 s 
process outbound tcp packet - accepted because of peer/peers: 64, ip: v4
                        time:   [1.2953 ms 1.3012 ms 1.3073 ms]
                        thrpt:  [76.493 Melem/s 76.854 Melem/s 77.202 Melem/s]
                 change:
                        time:   [-1.6444% -0.0915% +1.0298%] (p = 0.91 > 0.05)
                        thrpt:  [-1.0193% +0.0916% +1.6718%]
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
Benchmarking process outbound tcp packet - accepted because of peer/peers: 64, ip: v6: Collecting 100 samples in estimated 1.0642 s 
process outbound tcp packet - accepted because of peer/peers: 64, ip: v6
                        time:   [1.2884 ms 1.2929 ms 1.2978 ms]
                        thrpt:  [77.054 Melem/s 77.345 Melem/s 77.616 Melem/s]
                 change:
                        time:   [-10.330% -6.7017% -3.8547%] (p = 0.00 < 0.05)
                        thrpt:  [+4.0093% +7.1831% +11.520%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild

process inbound udp packet - rejected/peers: 64, ip: v4
                        time:   [3.4039 ms 3.4103 ms 3.4172 ms]
                        thrpt:  [29.264 Melem/s 29.323 Melem/s 29.378 Melem/s]
                 change:
                        time:   [-0.7512% -0.4361% -0.1216%] (p = 0.01 < 0.05)
                        thrpt:  [+0.1217% +0.4380% +0.7569%]
                        Change within noise threshold.
Found 17 outliers among 100 measurements (17.00%)
  4 (4.00%) high mild
  13 (13.00%) high severe
process inbound udp packet - rejected/peers: 64, ip: v6
                        time:   [4.1840 ms 4.1944 ms 4.2055 ms]
                        thrpt:  [23.778 Melem/s 23.842 Melem/s 23.901 Melem/s]
                 change:
                        time:   [-1.4360% -1.1389% -0.8479%] (p = 0.00 < 0.05)
                        thrpt:  [+0.8551% +1.1521% +1.4569%]
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

Benchmarking process inbound udp packet - accepted because of peer/peers: 64, ip: v4: Collecting 100 samples in estimated 1.0468 s (
process inbound udp packet - accepted because of peer/peers: 64, ip: v4
                        time:   [1.3325 ms 1.3387 ms 1.3452 ms]
                        thrpt:  [74.339 Melem/s 74.701 Melem/s 75.046 Melem/s]
                 change:
                        time:   [+0.3018% +0.8931% +1.5321%] (p = 0.01 < 0.05)
                        thrpt:  [-1.5090% -0.8852% -0.3009%]
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
Benchmarking process inbound udp packet - accepted because of peer/peers: 64, ip: v6: Collecting 100 samples in estimated 1.0514 s (
process inbound udp packet - accepted because of peer/peers: 64, ip: v6
                        time:   [1.3347 ms 1.3410 ms 1.3482 ms]
                        thrpt:  [74.171 Melem/s 74.569 Melem/s 74.925 Melem/s]
                 change:
                        time:   [+0.4559% +1.1328% +1.8404%] (p = 0.00 < 0.05)
                        thrpt:  [-1.8071% -1.1201% -0.4538%]
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

Benchmarking process inbound udp packet - accepted because of port/peers: 64, ip: v4: Collecting 100 samples in estimated 1.4681 s (
process inbound udp packet - accepted because of port/peers: 64, ip: v4
                        time:   [4.9272 ms 4.9383 ms 4.9501 ms]
                        thrpt:  [20.202 Melem/s 20.250 Melem/s 20.295 Melem/s]
                 change:
                        time:   [+0.4515% +0.7975% +1.1442%] (p = 0.00 < 0.05)
                        thrpt:  [-1.1313% -0.7912% -0.4494%]
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
Benchmarking process inbound udp packet - accepted because of port/peers: 64, ip: v6: Collecting 100 samples in estimated 1.1775 s (
process inbound udp packet - accepted because of port/peers: 64, ip: v6
                        time:   [5.8979 ms 5.9111 ms 5.9250 ms]
                        thrpt:  [16.878 Melem/s 16.917 Melem/s 16.955 Melem/s]
                 change:
                        time:   [+2.5654% +2.9485% +3.3182%] (p = 0.00 < 0.05)
                        thrpt:  [-3.2117% -2.8641% -2.5012%]
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

Benchmarking process outbound udp packet - accepted because of peer/peers: 64, ip: v4: Collecting 100 samples in estimated 1.0251 s 
process outbound udp packet - accepted because of peer/peers: 64, ip: v4
                        time:   [1.2845 ms 1.2909 ms 1.2986 ms]
                        thrpt:  [77.004 Melem/s 77.467 Melem/s 77.849 Melem/s]
                 change:
                        time:   [-2.1932% -1.5305% -0.8038%] (p = 0.00 < 0.05)
                        thrpt:  [+0.8103% +1.5543% +2.2424%]
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
Benchmarking process outbound udp packet - accepted because of peer/peers: 64, ip: v6: Collecting 100 samples in estimated 1.0411 s 
process outbound udp packet - accepted because of peer/peers: 64, ip: v6
                        time:   [1.3050 ms 1.3118 ms 1.3194 ms]
                        thrpt:  [75.791 Melem/s 76.229 Melem/s 76.629 Melem/s]
                 change:
                        time:   [-0.2107% +0.5178% +1.2853%] (p = 0.17 > 0.05)
                        thrpt:  [-1.2690% -0.5152% +0.2112%]
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

process outbound udp packet - accepted/peers: 64, ip: v4
                        time:   [5.6825 ms 5.6944 ms 5.7073 ms]
                        thrpt:  [17.521 Melem/s 17.561 Melem/s 17.598 Melem/s]
                 change:
                        time:   [-6.3597% -6.0450% -5.7097%] (p = 0.00 < 0.05)
                        thrpt:  [+6.0554% +6.4339% +6.7917%]
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
process outbound udp packet - accepted/peers: 64, ip: v6
                        time:   [6.7456 ms 6.7542 ms 6.7641 ms]
                        thrpt:  [14.784 Melem/s 14.806 Melem/s 14.825 Melem/s]
                 change:
                        time:   [-9.2827% -9.0358% -8.7875%] (p = 0.00 < 0.05)
                        thrpt:  [+9.6341% +9.9333% +10.233%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
  
  ## After
process inbound tcp packet - rejected/peers: 64, ip: v4
                        time:   [3.3899 ms 3.3973 ms 3.4055 ms]
                        thrpt:  [29.364 Melem/s 29.435 Melem/s 29.500 Melem/s]
                 change:
                        time:   [-0.0173% +0.3361% +0.7003%] (p = 0.06 > 0.05)
                        thrpt:  [-0.6954% -0.3350% +0.0173%]
                        No change in performance detected.
Found 17 outliers among 100 measurements (17.00%)
  4 (4.00%) high mild
  13 (13.00%) high severe
process inbound tcp packet - rejected/peers: 64, ip: v6
                        time:   [4.2277 ms 4.2420 ms 4.2577 ms]
                        thrpt:  [23.487 Melem/s 23.574 Melem/s 23.653 Melem/s]
                 change:
                        time:   [+0.2215% +0.6522% +1.1172%] (p = 0.00 < 0.05)
                        thrpt:  [-1.1048% -0.6479% -0.2210%]
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

Benchmarking process inbound tcp packet - accepted because of peer/peers: 64, ip: v4: Collecting 100 samples in estimated 1.0443 s (
process inbound tcp packet - accepted because of peer/peers: 64, ip: v4
                        time:   [1.3055 ms 1.3100 ms 1.3149 ms]
                        thrpt:  [76.053 Melem/s 76.338 Melem/s 76.598 Melem/s]
                 change:
                        time:   [-3.7117% -3.1809% -2.6376%] (p = 0.00 < 0.05)
                        thrpt:  [+2.7090% +3.2854% +3.8548%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
Benchmarking process inbound tcp packet - accepted because of peer/peers: 64, ip: v6: Collecting 100 samples in estimated 1.0525 s (
process inbound tcp packet - accepted because of peer/peers: 64, ip: v6
                        time:   [1.3556 ms 1.4143 ms 1.5132 ms]
                        thrpt:  [66.085 Melem/s 70.707 Melem/s 73.770 Melem/s]
                 change:
                        time:   [+1.2151% +5.5812% +14.384%] (p = 0.05 > 0.05)
                        thrpt:  [-12.575% -5.2862% -1.2005%]
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe

Benchmarking process inbound tcp packet - accepted because of port/peers: 64, ip: v4: Collecting 100 samples in estimated 1.4751 s (
process inbound tcp packet - accepted because of port/peers: 64, ip: v4
                        time:   [4.9668 ms 4.9832 ms 5.0006 ms]
                        thrpt:  [19.998 Melem/s 20.067 Melem/s 20.134 Melem/s]
                 change:
                        time:   [-20.959% -18.084% -16.363%] (p = 0.00 < 0.05)
                        thrpt:  [+19.565% +22.077% +26.517%]
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
Benchmarking process inbound tcp packet - accepted because of port/peers: 64, ip: v6: Collecting 100 samples in estimated 1.1735 s (
process inbound tcp packet - accepted because of port/peers: 64, ip: v6
                        time:   [5.8869 ms 5.8988 ms 5.9117 ms]
                        thrpt:  [16.916 Melem/s 16.953 Melem/s 16.987 Melem/s]
                 change:
                        time:   [-18.311% -17.815% -17.376%] (p = 0.00 < 0.05)
                        thrpt:  [+21.030% +21.677% +22.415%]
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

process outbound tcp packet - SYN/peers: 64, ip: v4
                        time:   [4.4186 ms 5.5524 ms 7.1576 ms]
                        thrpt:  [13.971 Melem/s 18.010 Melem/s 22.632 Melem/s]
                 change:
                        time:   [+5.5170% +32.691% +71.337%] (p = 0.02 < 0.05)
                        thrpt:  [-41.636% -24.637% -5.2285%]
                        Performance has regressed.
Found 13 outliers among 100 measurements (13.00%)
  4 (4.00%) high mild
  9 (9.00%) high severe
process outbound tcp packet - SYN/peers: 64, ip: v6
                        time:   [5.3924 ms 6.3669 ms 8.0233 ms]
                        thrpt:  [12.464 Melem/s 15.706 Melem/s 18.545 Melem/s]
                 change:
                        time:   [+5.1151% +24.445% +60.613%] (p = 0.04 < 0.05)
                        thrpt:  [-37.738% -19.643% -4.8662%]
                        Performance has regressed.
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) high mild
  10 (10.00%) high severe

Benchmarking process outbound tcp packet - other/peers: 64, ip: v4: Collecting 100 samples in estimated 1.1005 s (
process outbound tcp packet - other/peers: 64, ip: v4
                        time:   [2.6549 ms 2.6612 ms 2.6680 ms]
                        thrpt:  [37.481 Melem/s 37.577 Melem/s 37.667 Melem/s]
                 change:
                        time:   [+0.6339% +1.0850% +1.4905%] (p = 0.00 < 0.05)
                        thrpt:  [-1.4686% -1.0733% -0.6299%]
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
Benchmarking process outbound tcp packet - other/peers: 64, ip: v6: Collecting 100 samples in estimated 1.0026 s (
process outbound tcp packet - other/peers: 64, ip: v6
                        time:   [3.2326 ms 3.4903 ms 3.8853 ms]
                        thrpt:  [25.738 Melem/s 28.651 Melem/s 30.935 Melem/s]
                 change:
                        time:   [+1.8362% +9.9454% +22.190%] (p = 0.03 < 0.05)
                        thrpt:  [-18.160% -9.0458% -1.8031%]
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe

process outbound tcp packet - accepted because of peer/peers: 64, ip: v4
                        time:   [1.3061 ms 1.3127 ms 1.3196 ms]
                        thrpt:  [75.779 Melem/s 76.178 Melem/s 76.563 Melem/s]
                 change:
                        time:   [+0.1748% +0.8882% +1.6146%] (p = 0.01 < 0.05)
                        thrpt:  [-1.5890% -0.8804% -0.1745%]
                        Change within noise threshold.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
process outbound tcp packet - accepted because of peer/peers: 64, ip: v6
                        time:   [1.3042 ms 1.3104 ms 1.3170 ms]
                        thrpt:  [75.932 Melem/s 76.315 Melem/s 76.676 Melem/s]
                 change:
                        time:   [+0.7309% +1.3494% +1.9458%] (p = 0.00 < 0.05)
…